+++ src/error_chain.rs
@@ -1,12 +1,13 @@
 /// Prefer to use `error_chain` instead of this macro.
+#[doc(hidden)]
 #[macro_export]
-macro_rules! error_chain_processed {
+macro_rules! impl_error_chain_processed {
     // Default values for `types`.
     (
         types {}
         $( $rest: tt )*
     ) => {
-        error_chain_processed! {
+        impl_error_chain_processed! {
             types {
                 Error, ErrorKind, ResultExt, Result;
             }
@@ -21,7 +22,7 @@ macro_rules! error_chain_processed {
         }
         $( $rest: tt )*
     ) => {
-        error_chain_processed! {
+        impl_error_chain_processed! {
             types {
                 $error_name, $error_kind_name,
                 $result_ext_name;
@@ -391,7 +392,7 @@ macro_rules! error_chain_processing {
         }
     };
     ( ($a:tt, $b:tt, $c:tt, $d:tt) ) => {
-        error_chain_processed! {
+        impl_error_chain_processed! {
             types $a
             links $b
             foreign_links $c
@@ -400,8 +401,7 @@ macro_rules! error_chain_processing {
     };
 }
 
-/// This macro is used for handling of duplicated and out-of-order fields. For
-/// the exact rules, see `error_chain_processed`.
+/// Macro for generating error types and traits. See crate level documentation for details.
 #[macro_export]
 macro_rules! error_chain {
     ( $( $block_name:ident { $( $block_content:tt )* } )* ) => {